Only include .vue files in type def creation
The original logic for this build process walked the `src/components`
directory and created a type export for every file.  This was not taking
into account files that were not `.vue` single file components being put
there, which is now happening with the markdown example files.  This
ignores any files that do not end in `.vue` when generating types.